Split up Languages.kt for stronger buildSrc semantics
