+++ buildSrc/src/main/kotlin/configurations/Languages.kt
@@ -1,17 +1,12 @@
 package configurations
 
-import dependencies.Libraries.JUNIT_JUPITER_API
-import dependencies.Libraries.JUNIT_JUPITER_ENGINE
 import org.gradle.api.JavaVersion
 import org.gradle.api.Project
 import org.gradle.api.plugins.JavaPluginConvention
 import org.gradle.api.plugins.JavaPluginExtension
-import org.gradle.api.plugins.quality.Checkstyle
-import org.gradle.api.publish.PublishingExtension
-import org.gradle.api.publish.maven.MavenPublication
-import org.gradle.api.tasks.testing.Test
-
-import org.gradle.kotlin.dsl.*
+import org.gradle.kotlin.dsl.configure
+import org.gradle.kotlin.dsl.maven
+import org.gradle.kotlin.dsl.repositories
 
 object Languages {
     fun Project.attachLocalRepositories() {
@@ -39,41 +34,6 @@ object Languages {
         configure<JavaPluginExtension> {
             withJavadocJar()
             withSourcesJar()
-        }
-    }
-
-    fun Project.configureCheckstyle() {
-        tasks.withType<Checkstyle> {
-            configFile = file("$rootDir/gradle/lint/tnoodle-java.xml")
-        }
-    }
-
-    fun Project.configureJUnit5() {
-        dependencies {
-            add("testImplementation", JUNIT_JUPITER_API)
-            add("testRuntimeOnly", JUNIT_JUPITER_ENGINE)
-        }
-
-        tasks.withType<Test> {
-            useJUnitPlatform()
-
-            testLogging {
-                showStandardStreams = true
-            }
-        }
-    }
-
-    fun Project.configureMavenPublication(targetArtifactId: String? = null) {
-        configure<PublishingExtension> {
-            publications {
-                create<MavenPublication>("scrambler") {
-                    targetArtifactId?.let {
-                        artifactId = it
-                    }
-
-                    from(components["java"])
-                }
-            }
         }
     }
 }
