+++ sq12phase/build.gradle.kts
@@ -1,5 +1,5 @@
 import configurations.Languages.configureJava
-import configurations.Languages.configureMavenPublication
+import configurations.Publications.configureMavenPublication
 
 description = "A copy of Chen Shuang's square 1 two phase solver."
