+++ svglite/build.gradle.kts
@@ -1,7 +1,7 @@
 import configurations.Languages.attachRemoteRepositories
 import configurations.Languages.configureJava
-import configurations.Languages.configureCheckstyle
-import configurations.Languages.configureMavenPublication
+import configurations.Frameworks.configureCheckstyle
+import configurations.Publications.configureMavenPublication
 
 description = "A dead simple svg generation library written in pure Java, with no dependencies. This code runs on both desktop Java, Android, and compiles to Javascript with GWT."
